Forward port SVN commit 530022 by onurf:
Fixes drag and drop of messages from message tree to composer for creating attachments failing on 64-bit systems. 
   
Thanks to Andreas <apfaller@yahoo.com.au> for identifying the problem and providing the patch. 
     
CCBUG: 125447 


svn path=/trunk/KDE/kdepim/; revision=530041

+++ attachmentlistview.cpp
@@ -95,7 +95,7 @@ void AttachmentListView::contentsDropEvent( QDropEvent* e )
     QBuffer serNumBuffer( &serNums );
     serNumBuffer.open( QIODevice::ReadOnly );
     QDataStream serNumStream( &serNumBuffer );
-    unsigned long serNum;
+    quint32 serNum;
     KMFolder *folder = 0;
     int idx;
     QList<KMMsgBase*> messageList;
