+++ kmedit.cpp
@@ -153,7 +153,7 @@ void KMEdit::contentsDropEvent(QDropEvent *e)
         QBuffer serNumBuffer(&serNums);
         serNumBuffer.open(QIODevice::ReadOnly);
         QDataStream serNumStream(&serNumBuffer);
-        unsigned long serNum;
+        quint32 serNum;
         KMFolder *folder = 0;
         int idx;
         QList<KMMsgBase*> messageList;
